requirements.txt
apt-xapian-index==0.47
asgiref==3.2.7
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2020.4.5.1
chardet==3.0.4
command-not-found==0.3
conekta==2.4.0
cryptography==2.1.4
cupshelpers==1.0
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
Django==3.0.5
django-bootstrap-form==3.3
django-expiring-token==1.0.2
django-location-field==2.1.0
django-markdown-deux==1.0.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
email-reply-parser==0.5.9
gpg==1.10.0
httplib2==0.9.2
idna==2.9
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.1.1
markdown2==2.3.5
netifaces==0.10.4
nose==1.3.7
olefile==0.45.1
pbr==3.1.1
pexpect==4.2.1
Pillow==6.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
python-apt==1.6.5+ubuntu0.2
python-debian==0.1.32
python-magic==0.4.16
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
stevedore==1.27.1
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.25.9
vboxapi==1.0
virtualenv==16.7.7
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
whitenoise==3.3.1
xkit==0.0.0
